Develop

See merge request !70

+++ CGoGN/include/Geometry/tensor.hpp
@@ -23,6 +23,8 @@
  *******************************************************************************/
 
 #include <algorithm>
+#include <cmath>
+#include <cassert>
 
 namespace CGoGN {
 namespace Geom {
@@ -160,7 +162,7 @@ template <unsigned int SIZE, typename REAL>
 unsigned int
 Tensor<SIZE, REAL>::nbElem() const
 {
-	return pow(SIZE,m_order) ;
+	return (unsigned int)(pow(SIZE, m_order));
 }
 
 template <unsigned int SIZE, typename REAL>
@@ -184,7 +186,7 @@ template <unsigned int SIZE, typename REAL>
 bool
 Tensor<SIZE, REAL>::incremIndex(std::vector<unsigned int>& p)
 {
-	int i = p.size() - 1 ;
+	int i = int(p.size()) - 1 ;
 	while (i >= 0)
 	{
 		p[i] = (p[i] + 1) % SIZE ;
